Fixes #39403 - Moved all reusable logic from AreaChartLegend.js to .../helpers/LegendHelpers.js
