+++ webpack/assets/javascripts/react_app/components/common/charts/AreaChart/AreaChart.test.js
@@ -3,10 +3,10 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import AreaChart from './';
 import {
-  getXAxisTickValues,
   formatAxisTick,
   formatYAxisTick,
-} from './AreaChartLegend';
+  getXAxisTickValues,
+} from '../helpers/LegendHelpers';
 import { areaChartData, areaChartDataDenseSameMinute } from './AreaChart.fixtures';
 
 jest.unmock('./');
@@ -130,8 +130,8 @@ describe('AreaChart', () => {
   it('clicking legend symbol toggles series visibility', () => {
     const { container } = render(<AreaChart data={areaChartData.data} />);
 
-    const symbols = container.querySelectorAll('.area-chart-legend-symbol');
-    const labels = container.querySelectorAll('.area-chart-legend-label');
+    const symbols = container.querySelectorAll('.chart-legend-symbol');
+    const labels = container.querySelectorAll('.chart-legend-label');
 
     expect(symbols.length).toBeGreaterThanOrEqual(1);
     expect(labels.length).toBeGreaterThanOrEqual(1);
