+++ webpack/assets/javascripts/react_app/components/common/charts/helpers/LegendHelpers.js
@@ -1,31 +1,7 @@
 import React from 'react';
 import { ChartLabel, ChartPoint } from '@patternfly/react-charts';
 import { chart_color_black_500 as chartColorBlack500 } from '@patternfly/react-tokens';
-import './AreaChartLegend.scss';
-
-/** Process raw data into chart series. Backend may send Unix seconds or milliseconds. */
-export const processChartData = (data, xAxisDataLabel) => {
-  if (!data || data.length === 0) return null;
-  const timeColumn = data.find(col => col[0] === xAxisDataLabel);
-  if (!timeColumn) return null;
-  const timestamps = timeColumn.slice(1);
-  const toMs = val => {
-    const n = Number(val);
-    return n >= 1e12 ? n : n * 1000;
-  };
-  const dates = timestamps.map(t => new Date(toMs(t)));
-  const series = data
-    .filter(col => col[0] !== xAxisDataLabel)
-    .map(col => ({
-      name: col[0],
-      data: col.slice(1).map((value, index) => ({
-        x: dates[index],
-        y: value ?? 0,
-        name: col[0],
-      })),
-    }));
-  return series.length > 0 ? series : null;
-};
+import './LegendHelpers.scss';
 
 export const getSeriesOpacity = isDimmedByHover => {
   if (isDimmedByHover) return 0.35;
@@ -144,34 +120,6 @@ export const XAxisTickLabel = props => {
 };
 /* eslint-enable react/prop-types */
 
-/** Compute y-axis tick values from chart data. */
-export const getYTickValues = (chartData, config, hiddenSeries) => {
-  if (!chartData?.length) return undefined;
-  let maxY = 0;
-  if (config === 'timeseries') {
-    const n = chartData[0].data.length;
-    for (let i = 0; i < n; i++) {
-      let sum = 0;
-      chartData.forEach(chartSeries => {
-        if (!hiddenSeries.has(chartSeries.name))
-          sum += chartSeries.data[i]?.y ?? 0;
-      });
-      maxY = Math.max(maxY, sum);
-    }
-  } else {
-    chartData
-      .filter(s => !hiddenSeries.has(s.name))
-      .forEach(s => {
-        s.data.forEach(point => {
-          maxY = Math.max(maxY, point.y ?? 0);
-        });
-      });
-  }
-  if (maxY <= 0) return [0, 0.5, 1.0];
-  const step = Math.max(0.1, Math.ceil((maxY / 4) * 10) / 10);
-  return [0, 1, 2, 3, 4, 5].map(i => Math.round(i * step * 10) / 10);
-};
-
 export const formatAxisTick = t => {
   let date;
   if (t instanceof Date) {
@@ -232,7 +180,7 @@ export const InteractiveLegendSymbol = ({
   };
   return (
     <g
-      className="area-chart-legend-symbol"
+      className="chart-legend-symbol"
       onClick={handleClick}
       onKeyDown={e => e.key === 'Enter' && handleClick()}
       onMouseOver={() => datum?.name && setHoveredSeries(datum.name)}
@@ -260,7 +208,7 @@ export const InteractiveLegendLabel = ({
   };
   return (
     <g
-      className="area-chart-legend-label"
+      className="chart-legend-label"
       data-hidden={isHidden}
       data-hovered={isHovered}
       onClick={handleClick}
