+++ webpack/assets/javascripts/react_app/components/common/charts/helpers/LegendHelpers.scss
@@ -1,8 +1,8 @@
-.area-chart-legend-symbol {
+.chart-legend-symbol {
   cursor: pointer;
 }
 
-.area-chart-legend-label {
+.chart-legend-label {
   cursor: pointer;
 
   &[data-hidden='true'] {
